# Improve encoding/decoding speed