@Gabriel439 2019-10-31
@sjakobi: I did a code coverage check and the main things that are not exercised by the test suite are:

* Decoding the `IntegerToDouble`/`ListFold` built-ins
* Decoding large `Natural` numbers
* Encoding an empty list without a type annotation
* Decoding an integrity check
* Decoding an import beginning with `../` or `~/`
* Error messages

Only one of these concerns me: I believe the code is not correctly handling `Natural` numbers greater than `maxBound :: Word64`, which I'll fix