@sjakobi 2019-10-31
Thanks for checking! :) :+1: 

I can take care of adding test cases for the currently uncovered code to the test suite.